fix: resolve 14 bugs, architecture issues, and quality improvements
P0 bugs fixed:
- Remove duplicate CodeChunk from schema.py (dead code, real one in shared.py)
- Fix double-dot extension bug in config.py (.rb -> ..rb)
- Add effort_mode validation to 5 MCP thinking tools
- Fix __version__ to use importlib.metadata dynamic versioning

P1 architecture fixes:
- Delete 3000+ line session-ses_322e.md transcript, add session-*.md to .gitignore
- Split thinking_tools.py (2521 lines) into thinking_models.py, thinking_engine.py,
  and thinking_tools.py (MCP registration only) with backward-compatible re-exports
- Add MAX_SESSIONS=50 eviction for _large_write_buffers memory leak
- Add JSONL compaction for thinking memory to deduplicate strategy entries

P2 test coverage (318 tests, up from 280):
- Add test_mcp_wrappers.py: filesystem edge cases, Pydantic validation, error handling
- Add test_server.py: CLI arg parsing, _print_index_stats, SearchResultModel
- Add test_shared.py: embedder selection, context keys, CodeChunk fields

P3 code quality:
- Remove unused prek dependency from pyproject.toml
- Deduplicate dev deps (keep only [dependency-groups] dev)
- Wrap 9 blocking I/O tools in asyncio.to_thread()
