+++ pyproject.toml
@@ -31,16 +31,6 @@ dependencies = [
     "einops>=0.8.2",
 ]
 
-[project.optional-dependencies]
-dev = [
-    "pytest>=7.0.0",
-    "pytest-asyncio>=0.21.0",
-    "pytest-cov>=4.0.0",
-    "ruff>=0.1.0",
-    "mypy>=1.0.0",
-    "prek>=0.1.0",
-]
-
 [project.scripts]
 cocoindex-code = "cocoindex_code:main"
 
@@ -65,7 +55,6 @@ dev = [
     "pytest-cov>=4.0.0",
     "ruff>=0.1.0",
     "mypy>=1.0.0",
-    "prek>=0.1.0",
 ]
 
 [tool.uv]
