+++ src/cocoindex_code/__init__.py
@@ -1,7 +1,13 @@
 """CocoIndex Code - MCP server for indexing and querying codebases."""
 
+from importlib.metadata import PackageNotFoundError, version
+
 from .config import Config
 from .server import main, mcp
 
-__version__ = "0.1.0"
+try:
+    __version__ = version("cocoindex-code")
+except PackageNotFoundError:
+    __version__ = "0.0.0-dev"
+
 __all__ = ["Config", "main", "mcp"]
