+++ src/cocoindex_code/schema.py
@@ -1,20 +1,6 @@
 """Data models for CocoIndex Code."""
 
 from dataclasses import dataclass
-from typing import Any
-
-
-@dataclass
-class CodeChunk:
-    """Represents an indexed code chunk stored in SQLite."""
-
-    id: int
-    file_path: str
-    language: str
-    content: str
-    start_line: int
-    end_line: int
-    embedding: Any  # NDArray - type hint relaxed for compatibility
 
 
 @dataclass
